Added admin delete functionality
